Move the annotation to an api package
Move annotation and related public-facing classes to new package
Modify scope on non-public classes to help limit exposure
