+++ src/main/java/org/jboss/arquillian/testcontainers/TestContainersExtension.java
@@ -7,7 +7,7 @@ package org.jboss.arquillian.testcontainers;
 import org.jboss.arquillian.core.spi.LoadableExtension;
 import org.jboss.arquillian.test.spi.enricher.resource.ResourceProvider;
 
-public class TestContainersExtension implements LoadableExtension {
+class TestContainersExtension implements LoadableExtension {
     @Override
     public void register(ExtensionBuilder builder) {
         builder
